job_portal_fastapi/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
mysql-connector-python
pydantic-settings
passlib[bcrypt]
python-jose[cryptography]
python-multipart
alembic
fastapi-mail
